requirements.txt
python-telegram-bot==20.3
pycryptodome==3.18.0
